this should be a complete reversion of the last commit i did.
this is being done because after my last commit, i was promptly reminded
that kde is nearly a release so no changes besides bug fixes should be
done.

I have a copy of the newer sources in a seperate tree which will remain
uncommited until well after the 3.0 release.  i appologize for not thinking
ahead when commiting.

svn path=/trunk/kdeutils/kcalc/; revision=142652