Fix for bug number #206711. It's ugly but I see no other way to handle libgmp raising an abort.
So, We install a signal handler before the operation and if it fires, longjmp to the error handling case, finally
we restore the previous signal handler and proceed as normal.

Currently, this is all in an #ifdef Q_OS_LINUX since I have not tested on non-linux platforms, but I imagine that
it should in principle work on all platforms.

It is possible that other operations could also cause libgmp to try to allocate moew than INT_MAX bytes
so if this fix is considered the best approach, we may need to look into a more robust general "something
went wrong with libgmp" handler for all operation. Perhaps a thunk which wraps the real operations or
something.

BUG:206711


svn path=/trunk/KDE/kdeutils/kcalc/; revision=1167914