Do not apply style sheets when using default button colors
Qt cannot handle style sheets with a changed palette on "pixmap based"
styles, and falls back to the "Windows" style.

To work around styles' limitation, we only set the style sheet if any
of the colors has been customized. This ensures all buttons get a
consistent frame even if only one of the colors has been changed.

If you still see this bug, remove .kcalcrc

BUG: 237513
FIXED-IN: 4.6.0

svn path=/trunk/KDE/kdeutils/kcalc/; revision=1169124