making the const-ing style (left of type vs right of type) more consistent.
"const T &v" reads better, so I've decided to just use that accross the board.
I realize that with multiple levels of indirection the "T const *" is more
consistant (since all other levels of const must be on the right side)... but
we never use more than one level of indirection.


svn path=/trunk/KDE/kdeutils/kcalc/; revision=1171501