I believe that this patch should fix bug #267707.
Basically, I made KNumber when constructing from a QString take an optional
parameter which is to be used as the decimal separator.

The display object when dealing with a paste, will use the value of:

KGlobal::locale()->decimalSymbol()

for this parameter, so systems whose locale has the decimal separator set to ",", it should work ok.
thousands separator should not be an issue, since that is trimmed from the string before being passed.