Use math.h instead of cmath.
Also known as "let's try to prevent the #ifdef madness before it grows too
wild".

`isinf' and `isnan' are macros in C99, are not mentioned in C++03 and are
regular functions in C++11 (some C++ standard libraries such as GNU's
libstdc++ do implement those functions in C++98 and C++03 mode).

Simplify the current checks by unconditionally using math.h instead: this
allows us not to care about whether the `std' namespace is required and
assumes both `isinf' and `isnan' are macros, which are implemented when
undefined.

CCBUG:	320803
REVIEW:	111509