Fix invisible main window with Qt 5.6
Since Qt 5.6, the minimum size property seems to default to an empty size,
so it can't be used to set a fixed size for the main window.

A better approach is to tell the layout() to use a fixed size.

This change also exposed a redundant horizontal spacer in the .ui code, which
is now removed.

BUG: 360105
FIXED-IN: 16.04.1
REVIEW: 127847