Repeat last action when pressing "=" or Enter
This patch allows to press the equal sign again in order to repeat
the last action again.

For example "5+10" yields "15" on Enter and if you press Enter or
equal sign again it repeats the "+ 10" by showing you "25" now.

This behavior is disabled by default and can be turned on in
KCalc's settings panel under "Misc".

FEATURE: 108473

Differential Revision: https://phabricator.kde.org/D12099