Corrected behaviour on repeated pressing of operator buttons
Summary:
Fixes "wrong" behavior to the correct one as described in the bug report.
Before the fix pressing "10 * / 5 =" normally would result in 20 (Stack: (10,*), (10,/), (5,=)) but with those changes it results in 2 (Stack: (10, /), (5, =)).

Internally this is achieved by "blocking" node adding after a operator (*,-,/,...) has been pressed until a number is entered. If another operator is pressed whilst blocked the top (last) operator is changed instead.

BUG: 57349

Reviewers: cfeck, teran, #kde_applications

Reviewed By: cfeck

Subscribers: aacid, cfeck, kde-utils-devel

Differential Revision: https://phabricator.kde.org/D15463