Allow the string 'XOR' as a token for typed expressions
While ⊕ is used for the XOR pushbutton, most keyboard layouts cannot
type ⊕ directly so allowing the 'XOR' string is useful when typing
bitwise expressions. Using '^' would be more in line with '&' and '|'
but that's taken by the power operator already.

This particular comparison is case-insensitive because a user is
likely to type in lower case but the logical gate is properly
written in upper case so a case-insensitive comparison seems to
make the most sense