Show result while editing input
Show result while editing input without the need to press the equal
button, if the input reaches an invalid state (non computable) during
editing, result will be cleared or not depending on the error, inputs
that fail to be parsed like "1+#" trigger a result clear, inputs that
yield a math error like "1/0" also clear the result, however intermediate
inputs like "1+2+" will keep the "3" in the result display. If the user
attempts to commit an invalid input using the equal button, an error
message will appear as previously implemented. Changing the base mode
or the trigonometric mode also triggers a result recalculation and update.

A backspace signal clears the result display if the input reaches an
invalid state regardless of the error type.

Results are not committed to history unless the equal button is pressed.

Change default result display to be empty instead of "0", it will also
be empty if the input goes back to empty.

BUG:480607