Chain result upon equal clicked
Modify behavior to load last calculation result into the input display
when "equal" is clicked, this will also create a new entry in the history
panel. After loading the result this can be used for further calculations.

Also update behavior of "M+", "M-", "MS" and "Dat" to be compatible with
the new changes.

BUG: 487566